fix: store tracker_changeset_value_int value as BIGINT
Implements request #38647 store tracker_changeset_value_int value as BIGINT

How to test:
- before applying the patch, go to a tracker having an integer field
  allowing maximum 10+ characters.
- Create an artifact and set the value of the integer field to
  `9999999999`
- Submit
--> The integer field value is `2147483647`, which is the maximum
integer value the field can take, because it is stored on 4 bytes.

- apply the patch and `make dev-forgeupgrade`
- replay the steps 1 to 3
--> The integer field displays `9999999999`

+++ plugins/tracker/db/install.sql
@@ -335,7 +335,7 @@ CREATE TABLE tracker_changeset_value_file(
 DROP TABLE IF EXISTS tracker_changeset_value_int;
 CREATE TABLE tracker_changeset_value_int(
     changeset_value_id INT(11) NOT NULL PRIMARY KEY,
-    value INT(11) NULL
+    value BIGINT NULL
 ) ENGINE=InnoDB AUTO_INCREMENT=101;
 
 DROP TABLE IF EXISTS tracker_changeset_value_float;
